-Change vec/point/normal x/y/z accessors to be raw data member instead of function to match pbrt -Change onb to align to Frame in pbrt -Change ray to match pbrt with o and d data members -Add Transform::ApplyInverse function -Move OneMinusEpsilon from rng to mathinline -Add equality comparison operator for point2<T> -Add more numeric constants to mathinline -Add noise from pbrt to main (will probably remove) -Add SquareMatrix from pbrt to matrix.h -Add more pbrt math functions to mathinline -Add math.h from pbt for CatmullRomWeights (will probably remove) -Add HitP and InsideExclusive to Bounds

+++ src/sampler.cpp
@@ -1,5 +1,6 @@
 #include "sampler.h"
 #include "samplerBlueNoise.h"
+#include "mathinline.h"
 
 template <typename T>
 void Shuffle(T *samp, int count, int nDimensions, random_gen &rng) {
@@ -209,7 +210,7 @@ static inline vec2f sobol_calc_double_bluenoise(int  x, int  y, int i, int dim) 
 SobolSampler::SobolSampler(unsigned int maxSamples,
              random_gen& rng) : PixelSampler(1000000000,0,rng),
              current1Dsample(0), current2Dsample(0)  {
-  pixelseed = rng.unif_rand()*std::numeric_limits<unsigned int>::max();
+  pixelseed = rng.unif_rand()*(Float)std::numeric_limits<unsigned int>::max();
 }
 
 void SobolSampler::StartPixel( unsigned int i,  unsigned int j) {
